# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NayePankh Volunteer Hub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /volunteers:
    get:
      operationId: listVolunteers
      tags: [volunteers]
      summary: List all volunteers (admin/coordinator)
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 0 }
        - name: size
          in: query
          schema: { type: integer, default: 20 }
        - name: search
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string, enum: [ACTIVE, INACTIVE, PENDING] }
        - name: skill
          in: query
          schema: { type: string }
        - name: sortBy
          in: query
          schema: { type: string }
        - name: sortDir
          in: query
          schema: { type: string, enum: [asc, desc] }
      responses:
        "200":
          description: Page of volunteers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VolunteerPage"
  /volunteers/me:
    get:
      operationId: getMyProfile
      tags: [volunteers]
      summary: Get current volunteer's profile
      responses:
        "200":
          description: Volunteer profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VolunteerProfile"
    put:
      operationId: updateMyProfile
      tags: [volunteers]
      summary: Update current volunteer's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VolunteerProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VolunteerProfile"
  /volunteers/{volunteerId}:
    get:
      operationId: getVolunteerById
      tags: [volunteers]
      summary: Get a volunteer by ID
      parameters:
        - name: volunteerId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Volunteer profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VolunteerProfile"
        "404":
          description: Not found
...
  /volunteers/{volunteerId}/skills:
    post:
      operationId: addVolunteerSkill
      tags: [volunteers]
      summary: Add a skill to a volunteer
      parameters:
        - name: volunteerId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VolunteerSkillInput"
      responses:
        "200":
          description: Skills updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VolunteerProfile"
  /volunteers/{volunteerId}/skills/{skillId}:
    delete:
      operationId: removeVolunteerSkill
      tags: [volunteers]
      summary: Remove a skill from a volunteer
      parameters:
        - name: volunteerId
          in: path
          required: true
          schema: { type: integer }
        - name: skillId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Skill removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VolunteerProfile"

  # ── EVENTS ───────────────────────────────────────────────────────────────────
...
  /certificates:
    get:
      operationId: listCertificates
      tags: [certificates]
      summary: List certificates (own for volunteer, all for admin)
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 0 }
        - name: size
          in: query
          schema: { type: integer, default: 20 }
        - name: volunteerId
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Page of certificates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CertificatePage"
  /certificates/{certificateId}/download:
    get:
      operationId: downloadCertificate
      tags: [certificates]
      summary: Get certificate download URL
      parameters:
        - name: certificateId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Download URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CertificateDownload"
...
components:
  schemas:
...
    VolunteerProfile:
      type: object
      required: [id, email]
      properties:
        id:
          type: integer
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        name:
          type: ["string", "null"]
        email:
          type: string
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        college:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        age:
          type: ["integer", "null"]
        profileImage:
          type: ["string", "null"]
        availability:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        role:
          type: ["string", "null"]
        active:
          type: boolean
        badgeLevel:
          type: ["string", "null"]
          enum: [NONE, BRONZE, SILVER, GOLD, null]
        totalHours:
          type: number
        totalHoursLogged:
          type: ["number", "null"]
        eventsAttended:
          type: ["integer", "null"]
        eventsParticipated:
          type: ["integer", "null"]
        certificatesEarned:
          type: ["integer", "null"]
        skills:
          type: array
          items:
            $ref: "#/components/schemas/Skill"
        joinedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: ["string", "null"]
          format: date-time
        avatarUrl:
          type: ["string", "null"]
    VolunteerProfileUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        name:
          type: string
        phone:
          type: string
        address:
          type: string
        bio:
          type: string
        college:
          type: string
        city:
          type: string
        age:
          type: integer
        profileImage:
          type: string
        availability:
          type: string
    VolunteerSkillInput:
      type: object
      required: [skillId]
      properties:
        skillId:
          type: integer
    VolunteerPage:
      type: object
      required: [content, totalElements, totalPages, size, number]
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/VolunteerProfile"
        totalElements:
          type: integer
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer
        first:
          type: boolean
        last:
          type: boolean

    # ── EVENTS ──────────────────────────────────────────────────────────────────
...
    Certificate:
      type: object
      required: [id, volunteerId, eventId, issuedAt]
      properties:
        id:
          type: integer
        volunteerId:
          type: integer
        volunteerName:
          type: string
        eventId:
          type: integer
        eventTitle:
          type: string
        issuedAt:
          type: string
          format: date-time
        certificateNumber:
          type: string
        downloadUrl:
          type: ["string", "null"]
    CertificateDownload:
      type: object
      required: [url]
      properties:
        url:
          type: string
        expiresAt:
          type: ["string", "null"]
          format: date-time
    CertificatePage:
      type: object
      required: [content, totalElements, totalPages, size, number]
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/Certificate"
        totalElements:
          type: integer
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer
        first:
          type: boolean
        last:
          type: boolean

    # ── NOTIFICATIONS ─────────────────────────────────────────────────────────────
...
    Skill:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
...