# lib/api-zod/src/generated/types/volunteerProfile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NayePankh Volunteer Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { Skill } from './skill';
import type { VolunteerProfileBadgeLevel } from './volunteerProfileBadgeLevel';

export interface VolunteerProfile {
  id: number;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  name?: string | null;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  college?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  profileImage?: string | null;
  /** @nullable */
  availability?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  role?: string | null;
  active?: boolean;
  /** @nullable */
  badgeLevel?: VolunteerProfileBadgeLevel;
  totalHours?: number;
  /** @nullable */
  totalHoursLogged?: number | null;
  /** @nullable */
  eventsAttended?: number | null;
  /** @nullable */
  eventsParticipated?: number | null;
  /** @nullable */
  certificatesEarned?: number | null;
  skills?: Skill[];
  /** @nullable */
  joinedAt?: Date | null;
  /** @nullable */
  createdAt?: Date | null;
  /** @nullable */
  avatarUrl?: string | null;
}