# lib/api-zod/src/generated/types/volunteerProfileUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NayePankh Volunteer Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface VolunteerProfileUpdate {
  firstName?: string;
  lastName?: string;
  name?: string;
  phone?: string;
  address?: string;
  bio?: string;
  college?: string;
  city?: string;
  age?: number;
  profileImage?: string;
  availability?: string;
}